requirements.txt
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.1.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
jupyter>=1.0.0
jupyterlab>=3.4.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0
tensorflow>=2.9.0
torch>=1.12.0
imbalanced-learn>=0.9.0
optuna>=3.0.0
shap>=0.41.0
lime>=0.2.0.1
yellowbrick>=1.5
mlflow>=1.28.0
streamlit>=1.12.0
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0
python-dotenv>=0.20.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.64.0
joblib>=1.1.0
kaggle>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.3.0
celery>=5.2.0

# Development dependencies
pytest>=7.1.0
pytest-cov>=3.0.0
black>=22.6.0
flake8>=5.0.0
isort>=5.10.0
mypy>=0.971
pre-commit>=2.20.0

# Documentation
sphinx>=5.1.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0
